--- a/.gitignore
+++ b/.gitignore
@@ -8,7 +8,9 @@ tequilapi_spec.json
 /e2e/myst-provider/db
 tequilapi.json
 e2e/myst-provider/*.log.*
+e2e/prometheus/data
 mage_output_file.go
 coverage.txt
 localnet/volume/consumer
 localnet/volume/provider/db
+deployer