--- a/bin/docker/alpine-prebuilt/Dockerfile
+++ b/bin/docker/alpine-prebuilt/Dockerfile
@@ -1,4 +1,4 @@
-FROM alpine:3.6
+FROM alpine:3.11
 
 # Install packages
 RUN apk add --update --no-cache iptables ipset ca-certificates openvpn bash sudo openresolv