--- a/ci/test/e2e.go
+++ b/ci/test/e2e.go
@@ -18,7 +18,9 @@
 package test
 
 import (
+	"bytes"
 	"fmt"
+	"io/ioutil"
 	"os"
 
 	"github.com/magefile/mage/mg"
@@ -41,6 +43,11 @@ func BuildMystBinaryForE2eDocker() error {
 
 // BuildE2eTestBinary builds the e2e test binary.
 func BuildE2eTestBinary() error {
+	if err := replaceOpenvpnConnectionSetupPkg("github.com/mysteriumnetwork/go-openvpn/openvpn3", "github.com/mysteriumnetwork/node/mobile/mysterium/openvpn3"); err != nil {
+		return err
+	}
+	defer replaceOpenvpnConnectionSetupPkg("github.com/mysteriumnetwork/node/mobile/mysterium/openvpn3", "github.com/mysteriumnetwork/go-openvpn/openvpn3")
+
 	err := sh.RunWith(crossCompileFlags, "go", "test", "-c", "./e2e/")
 	if err != nil {
 		return err
@@ -59,7 +66,10 @@ func BuildE2eDeployerBinary() error {
 func TestE2EBasic() error {
 	logconfig.Bootstrap()
 
-	mg.Deps(BuildMystBinaryForE2eDocker, BuildE2eTestBinary, BuildE2eDeployerBinary)
+	mg.Deps(BuildMystBinaryForE2eDocker, BuildE2eDeployerBinary)
+
+	// not running this in parallel as it does some package switching magic
+	mg.Deps(BuildE2eTestBinary)
 
 	composeFiles := []string{
 		"./docker-compose.e2e-basic.yml",
@@ -119,4 +129,16 @@ func TestE2ECompatibility() error {
 		}
 	}
 	return nil
+}
+
+// replaceOpenvpnConnectionSetupPkg replaces openvpn_connection_setup.go go-openvpn pacakges
+// for mobile entry e2e tests so we don't need to include any C++ dependencies.
+func replaceOpenvpnConnectionSetupPkg(from, to string) error {
+	path := "./mobile/mysterium/openvpn_connection_setup.go"
+	content, err := ioutil.ReadFile(path)
+	if err != nil {
+		return err
+	}
+	content = bytes.Replace(content, []byte(from), []byte(to), 1)
+	return ioutil.WriteFile(path, content, 0600)
 }