--- a/e2e/gorunner/Dockerfile.precompiled
+++ b/e2e/gorunner/Dockerfile.precompiled
@@ -1,4 +1,4 @@
-FROM alpine:3.6
+FROM alpine:3.11
 
 RUN apk add --update --no-cache bash gcc musl-dev make linux-headers iptables ipset ca-certificates openvpn bash sudo openresolv