New:
 mage teste2ebasic  43.23s user 18.09s system 21% cpu **4:45.21** total

Old:
mage teste2ebasic  29.06s user 12.90s system 8% cpu **8:42.82** total

As a bonus, the mac now stays cool to the touch when running e2e. The savings will be smaller on linux, but I expect them to still be noticeable.